# openapi-task.yaml
openapi: 3.0.3
info:
  title: CS453 Final Task API
  version: 1.0.0
  description: First Swagger doc, describing basic CRUD operations on a list of tasks

security:
  - bearerAuth: []

servers:
  - url: http://localhost:3000

paths:
  /tasks:
    get:
      summary: Get list of tasks
      responses:
        "200":
          description: List of tasks returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: {"tasks": [{"id": "task-001","title": "Watch Week 6 lecture","course": "CS453","student_id": "djs001","completed": 0},{"id": "task-002","title": "Complete REST API exercise","course": "CS453","student_id": "djs001","completed": 1},{"id": "task-003","title": "Read authentication notes","course": "CS453","student_id": "student002","completed": 0}]}
    post:
      summary: "Adds new task to list"
      requestBody:
        description: "needs a title and status of task"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: "Null response"
        "400":
          description: "Invalid request, task not created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInvalidBody"

  /tasks/{id}:
    get:
      summary: Get task from list
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Numeric ID of the task to get
      responses:
        "200":
          description: Single task returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: "Resource requested not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNotFound"
    put:
      summary: Update one task in list
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the task to update
      requestBody:
        description: "needs title and status of task"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Single task returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: "Resource requested not found and could not be altered"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNotFound"
        "400":
          description: "Invalid request, task not updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInvalidBody"
    patch:
      summary: Update part of one task in list
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the task to update
      requestBody:
        description: "needs a combination of any of the following: title or status"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"

      responses:
        "200":
          description: Single task returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: "Resource requested not found and could not be altered"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNotFound"
        "400":
          description: "Invalid request, task not updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInvalidBody"
    delete:
      summary: Delete task from db
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the task to delete
      responses:
        "204":
          description: "Null response"
        "404":
          description: "Resource requested not found and therefore not deleted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorNotFound"


  #
  # Remember:
  # - Express routes use :id, such as /tasks/:id.
  # - OpenAPI paths use {id}, such as /tasks/{id}.
  #
  # You may define schemas under components/schemas if you want to avoid
  # repeating the same object structure many times.

components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - course
        - student_id
        - completed
      properties:
        id:
          type: string
          example: "task-001"
        title:
          type: string
          example: "Watch Week 6 lecture"
        course:
          type: string
          example: "CS453"
        student_id:
          type: string
          example: "djs001"
        completed:
          type: integer
          example: "0"
    TaskInput:
      type: object
      required:
        - id
        - title
        - course
        - student_id
        - completed
      properties:
        id:
          type: string
          example: "task-005"
        title:
          type: string
          example: "Watch Week 11 lecture"
        course:
          type: string
          example: "CS453"
        student_id:
          type: string
          example: "djs001"
        completed:
          type: integer
          example: "0"
  # DONE: Define an ErrorResponse schema with an error message.
    ErrorNotFound:
      type: object
      properties:
        Error:
          type: string
          example: "The resource you were looking for could not be found, please correct your path and try again"
    ErrorInvalidBody:
      type: object
      properties:
        Error:
          type: string
          example: "Invalid request body, please adjust your request and try again."
    UsernameAndPassword:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "user1@email.com"
        password:
          type: string
          example: "user-password"
    Username:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          example: "exampleuser"
  securitySchemes:
    bearerAuth:            # Arbitrary name for the scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # Optional hint for documentation